% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils1.R
\name{adjustable_leadtimes_robust}
\alias{adjustable_leadtimes_robust}
\title{Title}
\usage{
adjustable_leadtimes_robust(publication_months, valid_months)
}
\arguments{
\item{valid_months}{}
}
\description{
Title
}
\examples{
\dontrun{
valid_months_example <- c(4,5)
pub_months_example<- find_pub_mos(x =valid_months_example)
adjustable_leadtimes_robust(publication_months = pub_months_example,valid_months = valid_months_example)

}
}
