% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_weighted_forecast.R
\name{aggregate_weighted_forecast}
\alias{aggregate_weighted_forecast}
\title{aggregate_weighted_forecast}
\usage{
aggregate_weighted_forecast(df, df_area_loookup, analysis_level)
}
\arguments{
\item{df}{data.frame containing selected strata and forecasts data
temporally aggregated by selected publication and valid months}

\item{df_area_loookup}{data.frame containing selected strata and geographic
areas for weighted spatial aggregation.}

\item{analysis_level}{`character` vector of selected analysis level.}
}
\value{
data.frame with `value` column representing `weighted.mean` avg
    forecast value
}
\description{
aggregate_weighted_forecast
}
