% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{calc_weights}
\alias{calc_weights}
\title{Title}
\usage{
calc_weights(df, df_area_lookup = df_area_lookup, analysis_level)
}
\arguments{
\item{analysis_level}{}
}
\description{
Title
}
\examples{
silly string wrangling which should be removed by refactoring earlier
}
