% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_bucket_list.R
\name{dynamic_bucket_list}
\alias{dynamic_bucket_list}
\title{dynamic_bucket_list}
\usage{
dynamic_bucket_list(num_groups, reservoir)
}
\arguments{
\item{num_groups}{}
}
\description{
very clunky code to generate dynamic number of buckets for drag and drop.
Spent a long time trying to get it work with purrr which shoud have worked,
but I found myself in the depths of trying to debug rlang functions so I gave up.
purrr outputs were nearly working, but when passed to rlang::list2() produced and empty [[1]]
at the top that I couldn't get rid of.
}
\examples{
\dontrun{
dynamic_bucket_list("1")
dynamic_bucket_list("3")
}
}
