% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils1.R
\name{find_valid_month_interval}
\alias{find_valid_month_interval}
\title{find_latest_valid_month}
\usage{
find_valid_month_interval(valid_months)
}
\arguments{
\item{valid_months}{}
}
\value{
`integer` value of latest month
}
\description{
Correctly find latest month provided by user.
Especially useful if valid month window spans the new year Dec-Jan
}
\examples{
\dontrun{
# should throw these in testthat
find_valid_month_interval(valid_months = c(12,1,2,3,4,5))
find_valid_month_interval(valid_months = c(11,12,1,2,3,4,5))
find_valid_month_interval(valid_months = sort(c(11,12,1,2,3)))
find_valid_month_interval(valid_months = sort(c(12,1,2,3)))
find_valid_month_interval(valid_months = c(4,6,7))
}
}
