% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{get_spatial_filter_keys}
\alias{get_spatial_filter_keys}
\title{get_spatial_filter_keys}
\usage{
get_spatial_filter_keys(
  adm0_input,
  adm1_input,
  adm2_input,
  adm3_input,
  analysis_level
)
}
\arguments{
\item{adm3_input}{}
}
\description{
get_spatial_filter_keys
}
