% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_thresholding.R
\name{gt_style_thresh_table}
\alias{gt_style_thresh_table}
\title{gt_style_thresh_table}
\usage{
gt_style_thresh_table(gt_ob, table_type)
}
\arguments{
\item{gt_ob}{gt table class object}

\item{table_type}{`character` option to specify if table to be stylized is
the "overall_activation" or "strata" level table}
}
\value{
gt table object with stylings set
}
\description{
convenience function to stylize `{gt}` threshold table
}
