% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils1.R
\name{load_df_forecast}
\alias{load_df_forecast}
\title{load_df_forecast}
\usage{
load_df_forecast(dataset = "mars_lac")
}
\value{
`data.frame` with forecast in format required by app.
}
\description{
convenience function to load appropriate dataset for app development
Does need some work.
}
\examples{
\dontrun{
load_df_forecast(dataset= "mars_eth")
}
}
