% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_months_lts.R
\name{load_valid_mo_options}
\alias{load_valid_mo_options}
\title{load_valid_mo_options}
\usage{
load_valid_mo_options(valid_months)
}
\arguments{
\item{valid_months}{`integer` containeing valid month sequence}
}
\description{
load_valid_mo_options
}
