% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils1.R
\name{populate_admin_choices}
\alias{populate_admin_choices}
\title{admin_choices}
\usage{
populate_admin_choices(
  list_df = ldf,
  populate_with_analysis_level = "adm2_pcode",
  previous_selection_values = "ET06"
)
}
\arguments{
\item{admin_selection}{}
}
\description{
admin_choices
}
\examples{
\dontrun{
library(TriggerApp2024)
ldf$adm2
populate_admin_choices(list_df = ldf,
              populate_with_analysis_level = "adm2_pcode",
              previous_selection_values = "ET06"
              )
}
}
