% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_thresholding.R
\name{run_thresholding2}
\alias{run_thresholding2}
\title{same as below just df is already aggregated}
\usage{
run_thresholding2(df, leadtimes, analysis_level)
}
\arguments{
\item{df}{data.frame containing selected strata and forecasts data
temporally aggregated by selected publication and valid months}

\item{leadtimes}{selected leadtime values from leadtime sliders
returned from get_slider_values()  function.}

\item{analysis_level}{`character` vector of selected analysis level.}
}
\description{
same as below just df is already aggregated
}
