% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils1.R
\name{summarise_forecast_temporal}
\alias{summarise_forecast_temporal}
\title{summarise_forecast_temporal}
\usage{
summarise_forecast_temporal(df, publication_month, valid_month_arg)
}
\arguments{
\item{valid_month}{}
}
\description{
summarise_forecast_temporal
}
