% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{summarise_forecast_temporal2}
\alias{summarise_forecast_temporal2}
\title{summarise_forecast_temporal}
\usage{
summarise_forecast_temporal2(df, valid_month_arg)
}
\arguments{
\item{valid_month}{}
}
\description{
summarise_forecast_temporal
}
