% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_thresholding.R
\name{threshold_values}
\alias{threshold_values}
\title{threshold_values}
\usage{
threshold_values(df, slider_rps)
}
\arguments{
\item{slider_rps}{list of selected slider return period values.}
}
\value{
data.frame with theshold values for each strata and leadtime
    return period combination
}
\description{
Take return period values convert to quantiles and find quantile based
threshold on selected strata.
}
