% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils1.R
\name{ui_update_admin}
\alias{ui_update_admin}
\title{ui_update_admin}
\usage{
ui_update_admin(
  session = session,
  input = input,
  list_df,
  admin_level_choices = "adm2"
)
}
\arguments{
\item{admin_level_choices}{}
}
\description{
ui_update_admin
}
