% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{union_forecast_to_strata}
\alias{union_forecast_to_strata}
\title{Title}
\usage{
union_forecast_to_strata(df, df_area, analysis_level)
}
\arguments{
\item{df}{`data.frame` containing the seasonal/window sum for each leadtime and strata ... df_summarised <- summarise_forecast_temporal2(df, valid_month_arg)}
}
\description{
Title
}
