% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils1.R
\name{available_lts}
\alias{available_lts}
\title{available_lts}
\usage{
available_lts(publication_months, valid_months = c(5, 6))
}
\arguments{
\item{valid_months}{}
}
\value{
`numeric` named vector where value represent integer for month and name is the integer of leadtime
}
\description{
available_lts
}
\examples{
\dontrun{
valid_month_ex <- c(5,6)
available_lts(publication_months= find_pub_mos(valid_month_ex), valid_months=valid_month_ex)
available_lts(publication_months= 5, valid_months=valid_month_ex)
available_lts(publication_months= c(3,4), valid_months=4)
available_lts(publication_months= c(3), valid_months=4)
available_lts(publication_months= c(2), valid_months=4)

# what if there is a gap? should this not be allowed somehow or should it be handled
available_lts(publication_months= c(2,3),valid_months=c(4,6))
valid_mo_ex2 <- c(4,5)
all_pub_mo_ex2 <- find_pub_mos(valid_mo_ex2)
#> 11,12,1, 2, 3, 4 # looks good
 publication_months = all_pub_mo_ex2
available_lts(publication_months = c(12,1,2,3), valid_months= valid_mo_ex2)
}
}
