% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_months_lts.R
\name{find_month_range}
\alias{find_month_range}
\title{find_latest_valid_month}
\usage{
find_month_range(m)
}
\arguments{
\item{valid_months}{}
}
\value{
`integer` value of latest month
}
\description{
Correctly find latest month provided by user.
Especially useful if valid month window spans the new year Dec-Jan
}
\examples{
\dontrun{
# should throw these in testthat
find_valid_month_interval(valid_months = c(12,1,2,3,4,5))
find_month_range(m = c(11,12,1,2,3,4,5))
find_month_range(m = sort(c(11,12,1,2,3)))
find_month_range(m = sort(c(12,1,2,3)))
find_month_range(m = c(4,6,7))
find_month_range(m = c(10,11,12,1))
find_month_range(m = c(11,12,1,2,3))
find_month_range(m = c(1,2,3,11,12))
}
}
