% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils1.R
\name{find_pub_mos}
\alias{find_pub_mos}
\title{find_pub_mos}
\usage{
find_pub_mos(valid_months)
}
\arguments{
\item{valid_months}{`numeric` vector of valid months}
}
\value{
`integer` vector of publication months available based on supplied valid_months
}
\description{
find_pub_mos
}
\examples{
\dontrun{
find_pub_mos(valid_months = c(12,1,2,3,4,5))
find_pub_mos(valid_months = c(5,6))
}
}
