% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_df_forecast_parquets}
\alias{load_df_forecast_parquets}
\title{load_df_forecast_parquets}
\usage{
load_df_forecast_parquets(
  dataset_dir = ".data-scrap/orig_external_data",
  dataset = "mars_lac"
)
}
\value{
`data.frame` with forecast in format required by app.
}
\description{
convenience function to load appropriate dataset for app development
Does need some work.
}
\examples{
\dontrun{
load_df_forecast(dataset= "mars_eth")
}
}
