% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{run_thresholding2}
\alias{run_thresholding2}
\title{same as below just df is already aggregated}
\usage{
run_thresholding2(df, leadtimes, analysis_level)
}
\description{
same as below just df is already aggregated
}
